refactor: update import paths for BooksService and improve code formatting across components
