+++ phase-6/src/app/pages/home/top-bar/top-bar.component.ts
@@ -1,21 +1,19 @@
-import {Component, inject} from '@angular/core';
-import {FormsModule} from "@angular/forms";
-import {BooksService} from "../books.service";
+import { Component, inject } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { BooksService } from '../../../services/books.service';
 
 @Component({
   selector: 'app-top-bar',
   standalone: true,
-  imports: [
-    FormsModule
-  ],
+  imports: [FormsModule],
   templateUrl: './top-bar.component.html',
-  styleUrl: './top-bar.component.css'
+  styleUrl: './top-bar.component.css',
 })
 export class TopBarComponent {
   booksService: BooksService = inject(BooksService);
-  search = ''
+  search = '';
 
   searchBook() {
-    this.booksService.setSearchValue(this.search.toLowerCase())
+    this.booksService.setSearchValue(this.search.toLowerCase());
   }
 }
